redesign globalSearch
