--- src/components/ui/Button/Button.module.scss
@@ -1,9 +1,0 @@
-.button {
-  padding: 5px 10px;
-  background-color: $text-tertiary;
-  border-radius: 25px;
-  font-size: 18px;
-  font-weight: 700;
-  color: $text-secondary;
-  cursor: grabbing;
-}
