--- src/components/ui/Slider/Slider.module.scss
@@ -1,6 +1,0 @@
-.slider {
-  scroll-behavior: smooth;
-  display: flex;
-  gap: 10px;
-  white-space: nowrap;
-}
