--- src/routes/GlobalSearch/GlobalSearch.module.scss
@@ -1,55 +1,0 @@
-.search {
-  margin-top: 50px;
-  overflow-y: scroll;
-  max-height: calc(100vh - 160px - 100px);
-
-  &__entry {
-    display: flex;
-    align-items: center;
-  }
-
-  &__heading {
-    display: flex;
-    flex-direction: column;
-    gap: 50px;
-  }
-
-  &__title {
-    margin-left: 10px;
-
-    span {
-      color: $color-primary;
-    }
-  }
-
-  .ico__search {
-
-    width: 35px;
-    height: 35px;
-
-    path:nth-child(1),
-    path:nth-child(3) {
-      stroke: $color-primary;
-    }
-  }
-
-  &__items {
-    display: grid;
-    grid-template-columns: repeat(4, 1fr);
-    gap: 20px;
-    margin-top: 50px;
-  }
-}
-
-.search_error {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin: 0 auto;
-  height: calc(100vh - 160px - 100px);
-
-  &__title {
-    text-align: center;
-    margin: 0;
-  }
-}
