+++ src/routes/Watchlist.tsx
@@ -42,7 +42,7 @@ const Watchlist = () => {
     return (
       <section className="section-container">
         <h1 className="text-center text-4xl text-white font-bold mt-160">
-          Пока-что вы не добавили ни один фильм / сериал :)
+          Пока-что вы не добавили ни один фильм / сериал 🙁
         </h1>
       </section>
     );
